Added collidesWith() in Collider, removed items
collidesWith is now in Collider component to check for collisions. It
takes another Collider as a parameter and returns whether the two
Colliders overlapped. I also removed the Array<GameObject> items in
ShaftAction and replaced it with Array<Collider> and Array<ItemType>. I
figured this would be a bit easier for me instead of calling
getComponent() every frame.