Decrease speed of shaft when hitting "BOMB"
Shaft's speed now reverts to the base speed when hitting ItemType.BOMBs.
This is like the original game if I remember correctly. Achieved by
storing the speed set via setSpeed() and setting the speed to the stored
base speed whenever the shaft hit a BOMB. Added setSpeedIncrease as well
so I can replace that ugly 1.25f in ShaftAction.