Made by BOMBs disappear after set time
Before, BOMB-type items remain on the screen as long as they're not hit.
This lead to the inevitable point where all items are BOMBs and no other
type of item will spawn since maximum item count is already reached.
This situation basically ends the game. To prevent this, I added a
bombLife property to the ItemType component. By setting the bombLife via
setBombLife(), bombs  fadeOut() after the set amount of time.