Multiple item generation for ItemManager
Before, only one item is generated by the ItemManager every
coolDownTime. Because of this, there are many times when the player is
left staring at the screen with nothing to hit. Of course this id bad.
To remedy this, I made the ItemManager generate multiple items every
coolDownTime.
Instead of checking for the number of ACTIVE items, I checked if the
number of IDLE items is greater than itemCount - maxItemCount. Then, we
have a random number of items to generate. Finally, we fadeIn() those
items.